checkpoint

+++ CudaKeySearchDevice/CudaDeviceKeys.cuh
@@ -2,11 +2,12 @@
 #define _EC_CUH
 
 #include <cuda_runtime.h>
+#include "uint256.h"
 
 namespace ec {
-	__device__ unsigned int *getXPtr();
+	__device__ uint256 *getXPtr();
 
-	__device__ unsigned int *getYPtr();
+	__device__ uint256 *getYPtr();
 }
 
 #endif
